Refactor MailingLists::Subscribers to make it faster (#3128)
* Refactor MailingLists::Subscribers to make it faster (hitobito/hitobito_sac_cas#1470)

* Only show subscription_toggling button if list is subscribable

* Script to measure mailing_list subscriber performance:

```ruby
require "optparse"

options = {}
OptionParser.new do |opt|
  opt.on("--offset OFFSET") { |o| options[:offset] = o }
  opt.on("--limit LIMIT") { |o| options[:limit] = o }
end.parse!

def now = Process.clock_gettime(Process::CLOCK_MONOTONIC)

require_relative "config/environment"

branch = `git symbolic-ref HEAD | sed -e 's,.*/,,'`.strip
offset = options.fetch(:offset, 0).to_i
limit = options.fetch(:limit, 10).to_i
database = ActiveRecord::Base.connection.instance_variable_get(:@config)[:database]

batch_size = limit / 10
filename = "list-performance-#{database}-#{branch}-#{offset}-#{limit}.csv"

puts "Writing to #{filename}"
File.open(filename, "w") do |file|
  file.write CSV.generate_line(%w[id people seconds])
  MailingList.offset(offset).limit(limit).find_each.with_index do |list, index|
    puts "#{index} of #{limit}" if (index % batch_size) == 0
    file.flush if (index % batch_size) == 0
    start = now
    size = list.people.size
    file.write CSV.generate_line([list.id, size, (now - start).round(1)])
  end
end
```

---------

+++ app/helpers/mailing_lists_helper.rb
@@ -53,16 +53,20 @@ module MailingListsHelper
   private
 
   def button_unsubscribe
-    action_button(t("mailing_list_decorator.unsubscribe"),
-      group_mailing_list_subscription_path(@group, entry, entry.subscriptions.where(subscriber_id: current_user.id, subscriber_type: Person.sti_name).first.id),
-      :minus,
-      method: "delete")
+    if can?(:destroy, Subscription.new(mailing_list: entry, subscriber: current_user))
+      action_button(t("mailing_list_decorator.unsubscribe"),
+        group_person_subscription_path(entry.group, current_user, id: entry.id),
+        :minus,
+        method: "delete")
+    end
   end
 
   def button_subscribe
-    action_button(t("mailing_list_decorator.subscribe"),
-      group_mailing_list_person_index_path(@group, entry, subscription: {subscriber_id: current_user.id}),
-      :plus,
-      method: "post")
+    if can?(:create, Subscription.new(mailing_list: entry, subscriber: current_user))
+      action_button(t("mailing_list_decorator.subscribe"),
+        group_person_subscriptions_path(entry.group, current_user, id: entry.id),
+        :plus,
+        method: "post")
+    end
   end
 end
